==> analysis/01 - Generating Transactions From Charts.ipynb <==
# Generating Transactions From Charts
# By Gabriel P. Oliveira and Iago A. D. Vaz

import pandas as pd
import os
import csv
import ast

dataset_path = '../1. Collaboration Profiles in Music Genre Networks [@ ISMIR 2020]/dataset'

years = [2017, 2018, 2019]
markets = ['global', 'au', 'br', 'ca', 'de', 'fr', 'gb', 'jp', 'us']

# Getting artist info
artists = pd.read_csv(f'{dataset_path}/spotify_artists_info_complete_reduced_genres.csv', sep='\t', encoding='utf-8', index_col='artist_id')
artists.head()

# Getting hit songs info
hits = pd.read_csv(f'{dataset_path}/spotify_hits_dataset_complete.csv', sep='\t', encoding='utf-8', index_col='song_id')
hits.head()

# Getting charts
for market in markets:
    with open(f'dataset/charts/{market}-genre_transactions_charts.csv', 'w', encoding='utf-8') as outfile:
        writer = csv.writer(outfile, delimiter='\t')
        writer.writerow(['market', 'start_date', 'end_date', 'position', 'song_id', 'num_artists', 'genre_list', 'streams'])
        
        for year in years:
            chartdir = f'{dataset_path}/charts/{market}/{year}/with_features'
            charts = sorted(os.listdir(chartdir))
            
            for chartfile in charts:
                print(chartfile)
                with open(f'{chartdir}/{chartfile}', 'r', encoding='utf-8') as infile:
                    reader = csv.DictReader(infile, delimiter='\t')
                
                    # Getting start and end dates
                    chart = chartfile.replace(f'{market}-weekly_with_features-', '')
                    chart = chart.replace('.csv', '')
                    [start_date, end_date] = chart.split('--')

                    for row in reader:
                        sid = row['song_id']
                        
                        # Getting genre list from artists
                        genre_list = []
                        song_artists = ast.literal_eval(hits.loc[sid, 'artist_id'])
                        num_artists = hits.loc[sid, 'num_artists']
                        for aid in song_artists:
                            artist_genres = ast.literal_eval(artists.loc[aid, 'genres'])
                            genre_list += artist_genres 
                        
                        writer.writerow([market, start_date, end_date, row['position'], sid, num_artists, str(genre_list), row['streams']])
print('Finished.')
