==> analysis/02 - Aggregating Chart Positions.ipynb <==
# Aggregating Chart Positions
# By Gabriel P. Oliveira and Iago A. D. Vaz

import pandas as pd
import fim
import ast

markets = ['global', 'au', 'br', 'ca', 'de', 'fr', 'gb', 'jp', 'us']

for market in markets:
    hits = {}
    filename = f'dataset/charts/{market}-genre_transactions_charts.csv'
    print(filename)
    
    charts = pd.read_csv(filename, sep='\t', encoding='utf-8')
    
    for _, entry in charts.iterrows():
        sid = entry['song_id']
        
        if sid not in hits:
            hits[sid] = {
                'num_artists': entry['num_artists'], 
                'genre_list': entry['genre_list'],
                'top10_weeks': 0,
                'top30_weeks': 0,
                'top50_weeks': 0,
                'top100_weeks': 0,
                'top200_weeks': 0
            }
        
        if entry['position'] <= 10:
            hits[sid]['top10_weeks'] += 1
        elif entry['position'] <= 30:
            hits[sid]['top30_weeks'] += 1
        elif entry['position'] <= 50:
            hits[sid]['top50_weeks'] += 1
        elif entry['position'] <= 100:
            hits[sid]['top100_weeks'] += 1
        else:
            hits[sid]['top200_weeks'] += 1
        
    sorted_hits = {k: v for k, v in sorted(hits.items(), reverse=True, key=lambda item: item[1]['top200_weeks'])}
    sorted_hits = {k: v for k, v in sorted(sorted_hits.items(), reverse=True, key=lambda item: item[1]['top100_weeks'])}
    sorted_hits = {k: v for k, v in sorted(sorted_hits.items(), reverse=True, key=lambda item: item[1]['top50_weeks'])}
    sorted_hits = {k: v for k, v in sorted(sorted_hits.items(), reverse=True, key=lambda item: item[1]['top30_weeks'])}
    sorted_hits = {k: v for k, v in sorted(sorted_hits.items(), reverse=True, key=lambda item: item[1]['top10_weeks'])}

    hits_df = pd.DataFrame.from_dict(sorted_hits, orient='index')
    hits_df.to_csv(f'dataset/charts/{market}-genre_transactions_top_charts.csv', sep='\t', encoding='utf-8', index_label='song_id')
    
print('Finished.')

hits_df.head()
