==> analysis/04 - Mining Association Rules (RQ2).ipynb <==
# Mining Association Rules
# By Gabriel P. Oliveira and Iago A. D. Vaz

import pandas as pd
import fim
import ast

# Temporal

def temporal(market_, year_, drop_duplicate = False, target = 'c', minconf = 9, zmin = 2, report = 's'):
    charts = pd.read_csv(f'dataset/charts/{market_}-genre_transactions_charts.csv', sep='\t', encoding='utf-8')
    charts = charts.loc[charts['end_date'].str.contains(str(year_))]
    
    if drop_duplicate:
        charts.drop_duplicates(subset='song_id', keep='first', inplace=True)
    
    transactions = list([ast.literal_eval(x) for x in charts['genre_list']])
        
    return fim.apriori(transactions, target=target, conf=minconf, zmin=zmin, report=report)

markets = ['global', 'au', 'br', 'ca', 'de', 'fr', 'gb', 'jp', 'us']
years = [2017, 2018, 2019]
df_temporal = pd.DataFrame(columns=['consequent','antecedent', 'confidence', 'lift', 'year', 'market'])

target = 'r'
minconf = 20
zmin = 2
report = 'cl'
drop_duplicate = True
ascending = False


for market in markets:
    for year in years:
        freq_pat = temporal(market, year, drop_duplicate, target, minconf, zmin, report)
        df = pd.DataFrame(freq_pat, columns=[ 'consequent','antecedent', 'confidence', 'lift']).sort_values(by="lift", ascending=ascending).head(50)
        df['year'] = year
        df['market'] = market
        df_temporal = pd.concat([df_temporal, df])
        
df_temporal = df_temporal[['market', 'year','antecedent','consequent', 'confidence', 'lift']]

df_temporal.to_csv('dataset/assoc_rules_temporal.csv', index=False)

df_temporal.loc[df_temporal.market.isin(['br']) & (df_temporal.year == 2017)]

df_temporal.loc[df_temporal.market.isin(['br']) & (df_temporal.year == 2018)]

df_temporal.loc[df_temporal.market.isin(['br']) & (df_temporal.year == 2019)]

# Estratificada

def estratificada(market_, top_, drop_duplicate = False, target = 'c', minconf = 9, zmin = 2, report = 's'):
    charts = pd.read_csv(f'dataset/charts/{market_}-genre_transactions_charts.csv', sep='\t', encoding='utf-8')
    top_charts = pd.read_csv(f'dataset/charts/{market_}-genre_transactions_top_charts.csv', sep='\t', encoding='utf-8')
    
    charts = charts.merge(top_charts.drop(['num_artists', 'genre_list'], axis=1),on='song_id')
    
    if top_ == 10:
        charts = charts.loc[charts.top10_weeks > 0]
    elif top_ == 30:
        charts = charts.loc[(charts.top10_weeks == 0) & (charts.top30_weeks > 0)]
    elif top_ == 50:
        charts = charts.loc[(charts.top10_weeks == 0) & (charts.top30_weeks == 0) & (charts.top50_weeks > 0)]
    elif top_ == 100:
        charts = charts.loc[(charts.top10_weeks == 0) & (charts.top30_weeks == 0) & (charts.top50_weeks == 0) & (charts.top100_weeks > 0)]
    else: 
        charts = charts.loc[(charts.top10_weeks == 0) & (charts.top30_weeks == 0) & (charts.top50_weeks == 0) & (charts.top100_weeks == 0) & (charts.top200_weeks > 0)]
    
    if drop_duplicate:
        charts.drop_duplicates(subset='song_id', keep='first', inplace=True)
    
    transactions = list([ast.literal_eval(x) for x in charts['genre_list']])
        
    return fim.apriori(transactions, target=target, conf=minconf, zmin=zmin, report=report)

markets = ['global', 'au', 'br', 'ca', 'de', 'fr', 'gb', 'jp', 'us']
tops = [10, 30, 50, 100, 200]
df_estrat = pd.DataFrame(columns=['consequent','antecedent', 'confidence', 'lift', 'top', 'market'])

target = 'r'
minconf = 80
zmin = 2
report = 'cl'
drop_duplicate = True


for market in markets:
    for top in tops:
        freq_pat = estratificada(market, top, drop_duplicate, target, minconf, zmin, report)
        df = pd.DataFrame(freq_pat, columns=['consequent','antecedent', 'confidence', 'lift']).sort_values(by="lift", ascending=ascending).head(50)
        df['top'] = top
        df['market'] = market
        df_estrat = pd.concat([df_estrat, df])
        
df_estrat = df_estrat[['market', 'top','antecedent','consequent', 'confidence', 'lift']]

df_estrat.to_csv('dataset/assoc_rules_estrat.csv', index=False)

df_estrat.loc[df_estrat.market=='br']

# Parceria

def colab(market_, num_artists_, drop_duplicate = False, target = 'c', minconf = 9, zmin = 2, report = 's'):
    charts = pd.read_csv(f'dataset/charts/{market_}-genre_transactions_charts.csv', sep='\t', encoding='utf-8')
    
    charts = charts.loc[charts.num_artists >= num_artists_]
  
    
    if drop_duplicate:
        charts.drop_duplicates(subset='song_id', keep='first', inplace=True)
    
    transactions = list([ast.literal_eval(x) for x in charts['genre_list']])
        
    return fim.apriori(transactions, target=target, conf=minconf, zmin=zmin, report=report)

markets = ['global', 'au', 'br', 'ca', 'de', 'fr', 'gb', 'jp', 'us']
num_artists = [1, 2, 5]
df_colab = pd.DataFrame(columns=['consequent', 'antecedent', 'confidence', 'lift', 'colab_size', 'market'])

target = 'r'
minconf = 20
zmin = 2
report = 'cl'
drop_duplicate = True


for market in markets:
    for colab_size in num_artists:
        freq_pat = colab(market, colab_size, drop_duplicate, target, minconf, zmin, report)
        #df = pd.DataFrame(freq_pat, columns=['consequent', 'antecedent', 'confidence', 'lift']).sort_values(by="lift", ascending=ascending).head(50)
        df = pd.DataFrame(freq_pat, columns=['consequent', 'antecedent', 'confidence', 'lift']).sort_values(by="lift", ascending=ascending)
        df = df.loc[df.confidence < 0.5].head(10)
        
        df['colab_size'] = colab_size
        df['market'] = market
        df_colab = pd.concat([df_colab, df])
        
df_colab = df_colab[['market', 'colab_size','antecedent','consequent', 'confidence', 'lift']]

df_colab.to_csv('dataset/assoc_rules_colab.csv', index=False)

df_colab.loc[df_colab.market=='br']
