==> analysis/05 - Subgroup Discovery on Genre Networks (RQ3).ipynb <==
# Subgroup Discovery on Genre Networks
# By Gabriel P. Oliveira and Iago A. D. Vaz

import pandas as pd
import pysubgroup as ps
import numpy as np

dataset_path = '../1. Collaboration Profiles in Music Genre Networks [@ ISMIR 2020]/dataset'

market = 'us'
year = 2017

# Getting collaboration network
network = pd.read_csv(f'{dataset_path}/reduced_genre_network/{market}/{market}-reduced_genre_network-{year}.csv', sep=',', encoding='utf-8')
network.drop(['Avg_Popularity'], axis=1, inplace=True)
network.head()

# Getting collaboration profiles for each edge
profiles = pd.read_csv(f'{dataset_path}/reduced_genre_network/{market}/db_clusters/{market}_cluster_{year}.csv', sep='\t', encoding='utf-8')
profiles = profiles[['Source', 'Target', 'final.cluster']]
profiles['final.cluster'].replace({0: 'solid', 1: 'regular', 2: 'bridge', 4: 'emerging'}, inplace=True)
profiles.rename({'final.cluster': 'profile'}, axis=1, inplace=True)

profiles.set_index(['Source', 'Target'], inplace=True)
profiles.head()

# Adding collaboration profile in genre network
network['profile'] = np.nan

for idx, row in network.iterrows():
    if (row['Source'], row['Target']) in profiles.index:
        index = (row['Source'], row['Target'])
        network.loc[idx, 'profile'] = profiles.loc[index, 'profile']
    elif (row['Target'], row['Source']) in profiles.index:
        index = (row['Target'], row['Source'])
        network.loc[idx, 'profile'] = profiles.loc[index, 'profile']

network.head()

network['profile'].isna().sum()

# Getting node metrics
node_metrics = pd.read_csv(f'{dataset_path}/reduced_genre_network/{market}/reduced/nodes-{year}.csv', sep='\t', encoding='utf-8')
node_metrics.head()

# GENRE/NODE ATTRIBUTES
# Degree
# Weighted degree
# Closeness
# Betweenness
# Authority
# Hub
# Clustering Coefficient
# Pageranks

node_metrics = node_metrics[['Label', 'degree', 'weighted degree', 'closnesscentrality', 
                            'betweenesscentrality', 'Authority', 'pageranks', 'clustering']]

node_metrics.rename({'Label': 'genre', 
                     'closnesscentrality': 'closeness', 
                     'betweenesscentrality': 'betweenness', 
                     'Authority': 'authority',
                     'pageranks': 'pagerank'}, axis=1, inplace=True)

node_metrics.quantile([0, 0.5, 1])

# Putting node metrics into bins
attributes = ['degree', 'weighted degree', 'closeness', 'betweenness', 'authority', 'pagerank', 'clustering']

for att in attributes:
    node_metrics[att] = pd.qcut(node_metrics[att], q=[0, 0.5, 1], labels=['low', 'high'])
    #node_metrics[att] = pd.qcut(node_metrics[att], q=[0, 0.3, 0.8, 1], labels=['low', 'medium', 'high'])

node_metrics.set_index('genre', inplace=True)
node_metrics.head()

# Getting node metrics for node 1
network = network.join(node_metrics, on='Source', how='left')
network.rename({'degree': 'degree_1', 
                'weighted degree': 'wdegree_1', 
                'closeness': 'closeness_1', 
                'betweenness': 'betweenness_1', 
                'authority': 'authority_1',
                'hub': 'hub_1', 
                'pagerank': 'pagerank_1',
                'clustering': 'clustering_1'}, axis=1, inplace=True)
network.head()

# Getting node metrics for node 2
network = network.join(node_metrics, on='Target', how='left')
network.rename({'degree': 'degree_2', 
                'weighted degree': 'wdegree_2', 
                'closeness': 'closeness_2', 
                'betweenness': 'betweenness_2', 
                'authority': 'authority_2',
                'hub': 'hub_2', 
                'pagerank': 'pagerank_2',
                'clustering': 'clustering_2'}, axis=1, inplace=True)
network.head()

# Running Subgroup Discovery

target = ps.NumericTarget('Avg_Streams')

searchspace = ps.create_selectors(network, ignore=['Source', 'Target', 'Weight', 'Avg_Streams', 'wdegree_1', 'wdegree_2'])
task = ps.SubgroupDiscoveryTask(
    network,
    target,
    searchspace,
    result_set_size=5, # Number of subgroups returned
    depth=15, # Maximum amount of attributes
    qf=ps.StandardQFNumeric(a=1.0))

result = ps.BeamSearch().execute(task)
result_df = result.to_dataframe()

subgroups = []
for _, row in result_df.iterrows():
    # Print only subgroups with attributes on both nodes
    if '_1' in row.subgroup and '_2' in row.subgroup:
        subgroups.append(row.subgroup)
        print(f'{row.subgroup}')

# Getting metrics on subgroups
result_df[result_df['subgroup'].isin(subgroups)]

# Checking each subgroup
df = network[(network['closeness_2']=='high')
       & (network['degree_1']=='high')].sort_values(by=['Weight', 'Avg_Streams'], ascending=False)

print('Nodes:', len(set(list(df['Source']) + list(df['Target']))))
print('Edges:', len(df))

df['profile'].value_counts()

df
